Change the places of float and int

+++ python_variable_annotations/6-sum_mixed_list.py
@@ -5,7 +5,7 @@ Typing module is imported
 import typing
 
 
-def sum_mixed_list(mxd_lst: typing.List[typing.Union[float, int]]) -> float:
+def sum_mixed_list(mxd_lst: typing.List[typing.Union[int, float]]) -> float:
     """
     Function returns sum of floats and integers list
     """
